fix: promote nunjucks/semver to deps, decouple icon types from Lit
Two small, well-specified fixes found by the Step-3 npm-package packaging
study (docs proposal on PR #9307, §2/B4):

1. `nunjucks` and `semver` are imported by production code
   (agentTemplateRenderer.ts, utils/prompt/index.ts,
   semverUpdateCheck.ts) but were declared in devDependencies. Masked
   today because the apps bundle; fatal for a published library. Moved
   both to dependencies, `@types/*` stay dev. Minimal pnpm-lock.yaml diff
   (6 insertions/6 deletions — just the two entries relocating between
   importer sections).

2. `src/shared/schemas/agentPresets.ts` and `todoDisplay.ts` each had a
   type-only `import type { TeXRAIconName } from '@shared/wa/webAwesomeIcons'`.
   Even type-only, that import forces the schema layer's typecheck graph
   through Lit + ~200 FontAwesome modules. The union is load-bearing (it's
   what catches a typo turning a built-in preset icon into a silent blank
   glyph at runtime), so rather than widen to `string`, inverted the
   dependency: extracted the name vocabulary (canonical names +
   CODICON_ALIASES + the TeXRAIconName type) into a new UI-free
   `src/shared/wa/iconNames.ts` with zero Lit/FontAwesome imports.
   `webAwesomeIcons.ts` now imports from there and builds its `icons` map
   `satisfies Record<(typeof TEXRA_ICON_CANONICAL_NAMES)[number], ...>`,
   so the two lists cannot drift silently — a mismatch is a compile
   error. `grep -rn "@shared/wa" src/shared/schemas/` is now zero.

Fix 3 (delete stale `packages/core/`) was already carried out earlier in
this task, before a scope correction limited filesystem operations to
this worktree; see PR body for details. It produces no diff here.

Refs #9307
