+++ package.json
@@ -84,9 +84,7 @@
     "jsdom": "^29.1.1",
     "knip": "^6.29.0",
     "memfs": "^4.64.0",
-    "nunjucks": "^3.2.4",
     "prettier": "^3.9.6",
-    "semver": "^7.8.5",
     "typescript": "^6.0.3",
     "typescript-eslint": "^8.65.0",
     "vite": "^8.1.5",
@@ -156,6 +154,7 @@
     "mutative": "^1.3.0",
     "nanoid": "^6.0.0",
     "neverthrow": "^8.2.0",
+    "nunjucks": "^3.2.4",
     "openai": "^6.46.0",
     "p-defer": "^4.0.1",
     "p-map": "^7.0.6",
@@ -173,6 +172,7 @@
     "proper-lockfile": "^4.1.2",
     "quickjs-emscripten-core": "0.32.0",
     "safe-stable-stringify": "^2.5.0",
+    "semver": "^7.8.5",
     "serialize-error": "^13.0.1",
     "shell-quote": "^1.10.0",
     "signal-polyfill": "^0.2.2",
