+++ src/shared/wa/iconNames.ts
@@ -1,0 +1,247 @@
+/**
+ * The icon *name* vocabulary — UI-free.
+ *
+ * `webAwesomeIcons.ts` pairs each of these names with a Font Awesome icon
+ * definition (and imports Lit for the `waIcon` template helper), which pulls
+ * ~200 FontAwesome modules plus Lit into the type graph of anything that
+ * imports from it — even a bare `import type`. Consumers outside the UI layer
+ * (schemas, non-UI shared code) that only need the *name* union or the
+ * codicon-alias table should import from here instead.
+ *
+ * `webAwesomeIcons.ts` builds its `icons` map `satisfies
+ * Record<(typeof TEXRA_ICON_CANONICAL_NAMES)[number], FontAwesomeIconDefinition>`
+ * against {@link TEXRA_ICON_CANONICAL_NAMES}, so the two lists cannot drift:
+ * adding, removing, or renaming a canonical icon without updating the other
+ * file is a compile error, not a silent mismatch.
+ */
+
+// Canonical Font Awesome names, matching the keys of `icons` in
+// `webAwesomeIcons.ts` exactly (and in the same order). Add a new icon there
+// first, then mirror its key here — the `satisfies` constraint on `icons`
+// turns any mismatch into a type error.
+export const TEXRA_ICON_CANONICAL_NAMES = [
+  'arrow-down',
+  'arrow-left',
+  'arrow-right',
+  'arrow-rotate-left',
+  'arrow-up',
+  'arrow-up-right-from-square',
+  'arrows-rotate',
+  'backward-step',
+  'ban',
+  'bolt',
+  'book',
+  'bookmark',
+  'box',
+  'box-archive',
+  'brain',
+  'building',
+  'bullseye',
+  'caret-down',
+  'chart-line',
+  'chart-pie',
+  'check',
+  'check-double',
+  'chevron-down',
+  'chevron-left',
+  'chevron-right',
+  'chevron-up',
+  'circle',
+  'circle-check',
+  'circle-dot',
+  'circle-exclamation',
+  'circle-info',
+  'circle-question',
+  'circle-stop',
+  'circle-user',
+  'circle-xmark',
+  'clock',
+  'clock-rotate-left',
+  'cloud',
+  'cloud-arrow-down',
+  'cloud-arrow-up',
+  'code',
+  'code-branch',
+  'code-compare',
+  'code-merge',
+  'comment',
+  'comments',
+  'compass',
+  'compress',
+  'copy',
+  'cube',
+  'database',
+  'diagram-project',
+  'download',
+  'ellipsis',
+  'eraser',
+  'eye',
+  'file',
+  'file-circle-plus',
+  'file-code',
+  'file-export',
+  'file-lines',
+  'file-pdf',
+  'fish',
+  'flask',
+  'floppy-disk',
+  'folder',
+  'folder-open',
+  'folder-tree',
+  'forward-step',
+  'gear',
+  'gem',
+  'globe',
+  'graduation-cap',
+  'hashtag',
+  'heart',
+  'hexagon',
+  'image',
+  'key',
+  'lightbulb',
+  'link',
+  'list-check',
+  'list-ul',
+  'magnifying-glass',
+  'magnifying-glass-chart',
+  'meteor',
+  'microphone',
+  'minus',
+  'moon',
+  'note-sticky',
+  'palette',
+  'paper-plane',
+  'pencil',
+  'picture-in-picture',
+  'plane',
+  'play',
+  'plus',
+  'plus-minus',
+  'reply',
+  'right-from-bracket',
+  'right-to-bracket',
+  'robot',
+  'rocket',
+  'rotate-right',
+  'satellite',
+  'screwdriver-wrench',
+  'server',
+  'shield',
+  'spinner',
+  'star',
+  'terminal',
+  'thumbtack',
+  'thumbtack-slash',
+  'trash',
+  'triangle-exclamation',
+  'user',
+  'users',
+  'video',
+  'wand-magic-sparkles',
+  'window-maximize',
+  'wrench',
+  'xmark',
+] as const;
+
+type TeXRACanonicalIconName = (typeof TEXRA_ICON_CANONICAL_NAMES)[number];
+
+// Codicon-name aliases. Lets existing markup keep using familiar codicon names
+// (e.g. <wa-icon name="warning">) while resolving to the closest Font Awesome
+// glyph in the registry above. Prefer canonical names for new code.
+//
+// `as const satisfies …` keeps alias keys as a literal union (so
+// `keyof typeof CODICON_ALIASES` narrows TeXRAIconName) and compile-checks
+// every alias value against the canonical name list above.
+export const CODICON_ALIASES = Object.freeze({
+  account: 'circle-user',
+  add: 'plus',
+  archive: 'box-archive',
+  'arrow-small-down': 'caret-down',
+  beaker: 'flask',
+  'check-all': 'check-double',
+  checklist: 'list-check',
+  'circle-large-outline': 'circle',
+  'circle-outline': 'circle',
+  'circle-slash': 'circle-xmark',
+  'clear-all': 'eraser',
+  close: 'xmark',
+  'cloud-download': 'cloud-arrow-down',
+  'cloud-upload': 'cloud-arrow-up',
+  'comment-discussion': 'comments',
+  dash: 'minus',
+  'debug-continue': 'forward-step',
+  'debug-start': 'play',
+  'debug-stop': 'circle-stop',
+  'desktop-download': 'download',
+  'device-camera-video': 'video',
+  diff: 'code-compare',
+  'diff-added': 'plus',
+  // latexdiff marks additions/deletions — ± reads as that markup, where the
+  // previous not-equal (≠) glyph read as a math comparison.
+  'diff-multiple': 'plus-minus',
+  'diff-single': 'plus-minus',
+  discard: 'arrow-rotate-left',
+  edit: 'pencil',
+  error: 'circle-exclamation',
+  'file-media': 'image',
+  files: 'copy',
+  fold: 'compress',
+  'folder-library': 'folder-tree',
+  'folder-opened': 'folder-open',
+  github: 'code-branch',
+  'git-commit': 'circle-dot',
+  'git-merge': 'code-merge',
+  graph: 'chart-line',
+  'graph-line': 'chart-line',
+  history: 'clock-rotate-left',
+  info: 'circle-info',
+  inspect: 'magnifying-glass-chart',
+  library: 'book',
+  'list-tree': 'list-ul',
+  loading: 'spinner',
+  merge: 'code-merge',
+  mic: 'microphone',
+  'mortar-board': 'graduation-cap',
+  'new-file': 'file-circle-plus',
+  note: 'note-sticky',
+  organization: 'building',
+  output: 'terminal',
+  package: 'box',
+  'pass-filled': 'circle-check',
+  'pie-chart': 'chart-pie',
+  pulse: 'chart-line',
+  question: 'circle-question',
+  references: 'link',
+  refresh: 'rotate-right',
+  save: 'floppy-disk',
+  search: 'magnifying-glass',
+  send: 'paper-plane',
+  'server-process': 'server',
+  'settings-gear': 'gear',
+  'sign-in': 'right-to-bracket',
+  'sign-out': 'right-from-bracket',
+  'source-control': 'code-branch',
+  sparkle: 'wand-magic-sparkles',
+  'stop-circle': 'circle-stop',
+  stylesheet: 'palette',
+  'symbol-method': 'cube',
+  'symbol-number': 'hashtag',
+  'symbol-numeric': 'hashtag',
+  'symbol-operator': 'cube',
+  'symbol-structure': 'diagram-project',
+  sync: 'arrows-rotate',
+  tasklist: 'list-check',
+  tools: 'screwdriver-wrench',
+  'type-hierarchy': 'diagram-project',
+  wand: 'wand-magic-sparkles',
+  warning: 'triangle-exclamation',
+  window: 'window-maximize',
+  x: 'xmark',
+  zap: 'bolt',
+} as const satisfies Readonly<Record<string, TeXRACanonicalIconName>>);
+
+// Names accepted by <wa-icon library="texra" name="..."> — canonical names plus
+// the codicon-style aliases. Exported so callers can type-check icon usage
+// without importing Lit/FontAwesome (see module doc comment above).
+export type TeXRAIconName =
+  TeXRACanonicalIconName | keyof typeof CODICON_ALIASES;
